handle error functions moved to utils from controllers
